Fix cannot mapped
Fix cannot mapped  and change id type from int to long

+++ RawMaterials/Models/Entities/Advertizment.cs
@@ -11,7 +11,7 @@ namespace RawMaterials.Models.Entities
 
         public TeamWork TeamWork { get; set; }
 
-        public int TeamWorkId { get; set; }
+        public long TeamWorkId { get; set; }
 
         public PaymentInfo PaymentInfo { get; set; }
