+++ RawMaterials/Models/Entities/City.cs
@@ -8,14 +8,14 @@ namespace RawMaterials.Models.Entities
 {
     public class City
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Name { get; set; }
 
         public Province Province { get; set; }
 
-        public int ProvinceId { get; set; }
-
+        public long ProvinceId { get; set; }
+     
 
     }
 }
