+++ RawMaterials/Models/Entities/ImporterCategory.cs
@@ -15,7 +15,7 @@ namespace RawMaterials.Models.Entities
 
         public Importer Importer { get; set; }
 
-        public int ImporterId { get; set; }
+        public long ImporterId { get; set; }
