+++ RawMaterials/Models/Entities/InterestMaterial.cs
@@ -15,6 +15,6 @@ namespace RawMaterials.Models.Entities
 
         public Importer Importer { get; set; }
 
-        public int ImporterId { get; set; }
+        public long ImporterId { get; set; }
     }
 }
