+++ RawMaterials/Models/Entities/SuplierCategory.cs
@@ -15,6 +15,6 @@ namespace RawMaterials.Models.Entities
 
         public Suplier Suplier { get; set; }
 
-        public int SuplierId { get; set; }
+        public long SuplierId { get; set; }
     }
 }
